[refactor] coverLetter api & hooks 위치 shared로 통일
