+++ frontend/src/shared/hooks/useCoverLetterQueries.ts
@@ -1,9 +1,10 @@
-import { useCallback } from 'react';
+import { useCallback, useContext } from 'react';
 
 import {
   useMutation,
   useQuery,
   useQueryClient,
+  useSuspenseInfiniteQuery,
   useSuspenseQueries,
   useSuspenseQuery,
 } from '@tanstack/react-query';
@@ -11,9 +12,14 @@ import {
 import {
   createCoverLetter,
   deleteCoverLetter,
+  fetchScraps,
+  fetchSharedLink,
   getCoverLetter,
+  searchCoverLetters,
+  toggleSharedLinkStatus,
 } from '@/shared/api/coverLetterApi';
 import { getQnAIdList } from '@/shared/api/qnaApi';
+import { ToastMessageContext } from '@/shared/context/ToastMessageContext';
 import { coverLetterQueryKeys } from '@/shared/hooks/queries/coverLetterKeys';
 import { homeKeys } from '@/shared/hooks/queries/homeKeys';
 import { libraryKeys } from '@/shared/hooks/queries/libraryKeys';
@@ -122,5 +128,96 @@ export const useDeleteCoverLetter = () => {
   return useMutation<void, Error, { coverLetterId: number }>({
     mutationFn: (variables) => deleteCoverLetter(variables.coverLetterId),
     onSuccess: () => invalidate(),
+  });
+};
+
+// 기업명 직무명 기반 검색
+export const useCoverLetterSearch = (searchWord = '', size = 9, page = 1) => {
+  return useSuspenseQuery({
+    queryKey: ['coverletter', 'search', { searchWord, size, page }],
+    queryFn: () => searchCoverLetters({ searchWord, size, page }),
+    staleTime: 5 * 60 * 1000,
+  });
+};
+
+// 스크랩한 자기소개서 내에서 검색 (무한 스크롤)
+export const useScrapCoverLetters = (searchWord = '', size = 9) => {
+  return useSuspenseInfiniteQuery({
+    queryKey: ['coverletter', 'scrap', { searchWord, size }],
+    queryFn: ({ pageParam }) =>
+      fetchScraps({ searchWord, size, lastQnaId: pageParam }),
+    initialPageParam: undefined as number | undefined,
+    getNextPageParam: (lastPage) => {
+      const lastScrap = lastPage.scraps.at(-1);
+      if (!lastPage.hasNext || !lastScrap) return undefined;
+      return lastScrap.id;
+    },
+    staleTime: 5 * 60 * 1000,
+  });
+};
+
+// 첨삭 링크 조회
+export const useSharedLink = (
+  coverLetterId: number,
+  enabled: boolean = true,
+) => {
+  return useQuery({
+    queryKey: ['coverletter', 'sharedLink', { coverLetterId }],
+    queryFn: () => fetchSharedLink({ coverLetterId }),
+    enabled,
+  });
+};
+
+// 첨삭 링크 활성화/비활성화
+export const useSharedLinkToggle = () => {
+  const queryClient = useQueryClient();
+  const toast = useContext(ToastMessageContext);
+
+  return useMutation({
+    mutationFn: ({
+      coverLetterId,
+      active,
+    }: {
+      coverLetterId: number;
+      active: boolean;
+    }) => toggleSharedLinkStatus({ coverLetterId, active }),
+    onSuccess: (_, { coverLetterId, active }) => {
+      queryClient.invalidateQueries({
+        queryKey: ['coverletter', 'sharedLink', { coverLetterId }],
+      });
+
+      if (active) {
+        queryClient.invalidateQueries({
+          queryKey: ['reviews', { coverLetterId }],
+        });
+      } else {
+        queryClient.removeQueries({
+          queryKey: ['reviews', { coverLetterId }],
+        });
+      }
+
+      const qnaIdList = queryClient.getQueryData<number[]>([
+        'qnaIdList',
+        coverLetterId,
+      ]);
+
+      if (qnaIdList) {
+        queryClient.invalidateQueries({
+          queryKey: ['qnaIdList', coverLetterId],
+        });
+      }
+
+      queryClient.invalidateQueries({ queryKey: ['qna'] });
+
+      toast?.showToast('첨삭 링크 상태가 변경되었습니다.', true);
+    },
+    onError: (error: unknown) => {
+      const message =
+        error instanceof Error
+          ? error.message
+          : '첨삭 링크 상태 변경에 실패했습니다.';
+
+      toast?.showToast(message, false);
+    },
   });
 };
